+++ src/Helper.cpp
@@ -7,7 +7,7 @@ void NormalizeString(std::string* string) {
 			if(!(c == ' ' || c == '\t')){ break; }
 			string->erase(string->begin());
 		}
-	for(int i = string->length() - 1; i > 0; i--){
+	for(size_t i = string->length() - 1; i > 0; i--){
 		char c = string->at(i);
 		if(!(c == ' ' || c == '\t')){ break; }
 		string->pop_back();
